image_fraud/app/requirements.txt
matplotlib==3.8.3
numpy==1.26.4
opencv_python==4.9.0.80
Pillow==10.3.0
psycopg2==2.9.9
scikit_learn==1.3.2
scipy==1.13.1
tensorboardX==2.6.2.2
tensorboardX==2.6.2.2
torch==2.0.0
torchvision==0.15.1
tqdm==4.66.2
python-decouple==3.8